requirements.txt
pypdf2
sentence-transformers
faiss-cpu
transformers
torch
numpy
tabulate
